requirements.txt
PySide6>=6.4.0
pywin32>=305
mss>=9.0.1
volcengine>=1.0.0
requests>=2.28.0
python-dotenv>=0.21.0
pillow>=9.3.0